requirements.txt
psycopg2-binary==2.7.5
pyModbusTCP==0.1.8
